// frontend/src/api/models/billDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { MemberDto } from "./memberDto";
import type { ExpenseDto } from "./expenseDto";

export interface BillDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  shareCode?: string | null;
  version?: number;
  isSettled?: boolean;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  members?: MemberDto[] | null;
  /** @nullable */
  expenses?: ExpenseDto[] | null;
  /** @nullable */
  settledTransfers?: string[] | null;
}